reorganise installables modules
