+++ src/cloudai/_core/installables/__init__.py
@@ -1,0 +1,32 @@
+# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
+# Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .base import Installable, InstallStatusResult
+from .docker_image import DockerImage
+from .file import File
+from .git_repo import GitRepo
+from .hf_model import HFModel
+from .python_executable import PythonExecutable
+
+__all__ = [
+    "DockerImage",
+    "File",
+    "GitRepo",
+    "HFModel",
+    "InstallStatusResult",
+    "Installable",
+    "PythonExecutable",
+]
