+++ tests/test_hf_model_manager.py
@@ -36,17 +36,14 @@ def installer(slurm_system, tmp_path: Path) -> BaseInstaller:
 def test_download(hf_model: HFModel, installer: BaseInstaller) -> None:
     assert hf_model._installed_path is None
 
-    with patch("cloudai._core.installables.huggingface_hub.snapshot_download", return_value=str("/real/path")):
+    with patch("huggingface_hub.snapshot_download", return_value=str("/real/path")):
         hf_model.install(installer)
 
     assert hf_model.installed_path == Path("/real/path")
 
 
 def test_download_raises(hf_model: HFModel, installer: BaseInstaller) -> None:
-    with patch(
-        "cloudai._core.installables.huggingface_hub.snapshot_download",
-        side_effect=Exception("some error message"),
-    ):
+    with patch("huggingface_hub.snapshot_download", side_effect=Exception("some error message")):
         result = hf_model.install(installer)
 
     assert not result.success
@@ -55,7 +52,7 @@ def test_download_raises(hf_model: HFModel, installer: BaseInstaller) -> None:
 
 
 def test_is_downloaded(hf_model: HFModel, installer: BaseInstaller) -> None:
-    with patch("cloudai._core.installables.huggingface_hub.snapshot_download", return_value=str("/real/path")):
+    with patch("huggingface_hub.snapshot_download", return_value=str("/real/path")):
         result = hf_model.is_installed(installer)
 
     assert result.success
@@ -64,10 +61,7 @@ def test_is_downloaded(hf_model: HFModel, installer: BaseInstaller) -> None:
 
 
 def test_is_downloaded_raises(hf_model: HFModel, installer: BaseInstaller) -> None:
-    with patch(
-        "cloudai._core.installables.huggingface_hub.snapshot_download",
-        side_effect=Exception("some error message"),
-    ):
+    with patch("huggingface_hub.snapshot_download", side_effect=Exception("some error message")):
         result = hf_model.is_installed(installer)
 
     assert not result.success
